Add infinite scroll to manage page
