+++ src/app/(admin)/articles/manage/clientPage.tsx
@@ -1,23 +1,19 @@
 "use client";
 
-import { useEffect, useRef, useState } from "react";
-import { listPublishedArticles } from "@/services/article-api";
-import { mapApiArticleToArticle } from "@/lib/type-mappers";
-import { Article, ArticleFilter, ArticleSortBy } from "@/types/article";
 import ArticleList from "@/components/Article/ArticleList";
+import { listArticles } from "@/services/article-api";
+import { ArticleFilter, Article, ArticleSortBy } from "@/types/article";
+import { mapApiArticleToArticle } from "@/lib/type-mappers";
+import { useEffect, useState, useRef } from "react";
 
-export interface ClientPageProps {
-  selectedTagId?: string;
-}
-
-export default function ClientPage({ selectedTagId }: ClientPageProps) {
+export default function ClientPage() {
   const [articles, setArticles] = useState<Article[]>([]);
   const [sortBy, setSortBy] = useState<ArticleSortBy>(ArticleSortBy.Latest);
-
+  const [onlyDrafts, setOnlyDrafts] = useState(false);
+  const [onlyArchived, setOnlyArchived] = useState(false);
   const [currentPage, setCurrentPage] = useState<number>(1);
   const [lastPageReached, setLastPageReached] = useState<boolean>(false);
-
-  const [loading, setLoading] = useState(true);
+  const [loading, setLoading] = useState<boolean>(true);
   const [loadingMore, setLoadingMore] = useState(false);
   const [error, setError] = useState<string | null>(null);
   const loaderRef = useRef<HTMLDivElement | null>(null);
@@ -30,33 +26,37 @@ export default function ClientPage({ selectedTagId }: ClientPageProps) {
     setLoading(true);
     setLoadingMore(false);
     setError(null);
-  }, [selectedTagId, sortBy]);
+  }, [sortBy, onlyDrafts, onlyArchived]);
 
   useEffect(() => {
+    let statuses: string[] = [];
+    if (onlyDrafts && !onlyArchived) {
+      statuses = ["Draft"];
+    } else if (onlyArchived && !onlyDrafts) {
+      statuses = ["Archived"];
+    }
     const filter: ArticleFilter = {
       Keyword: "",
-      Statuses: [],
-      TagIds: selectedTagId ? [selectedTagId] : [],
+      Statuses: statuses,
+      TagIds: [],
       SortBy: sortBy.toString(),
       Page: currentPage,
       PageSize: 10,
     };
 
     let isCancelled = false;
-    listPublishedArticles(filter)
+    listArticles(filter)
       .then((apiArticles) => {
         if (isCancelled || !apiArticles) return;
         setArticles((prev) => {
           const mapped = apiArticles.results.map(mapApiArticleToArticle);
-          // Replace on first page, append on subsequent pages
           const next = filter.Page === 1 ? mapped : prev.concat(mapped);
-          // Set lastPageReached if all items loaded
           if (next.length >= apiArticles.totalCount) setLastPageReached(true);
           return next;
         });
       })
       .catch(() => {
-        if (!isCancelled) setError("Failed to load content.");
+        if (!isCancelled) setError("Failed to load articles.");
       })
       .finally(() => {
         if (!isCancelled) {
@@ -67,7 +67,7 @@ export default function ClientPage({ selectedTagId }: ClientPageProps) {
     return () => {
       isCancelled = true;
     };
-  }, [currentPage, selectedTagId, sortBy]);
+  }, [currentPage, sortBy, onlyDrafts, onlyArchived]);
 
   useEffect(() => {
     if (loading || loadingMore || lastPageReached) return;
@@ -89,13 +89,17 @@ export default function ClientPage({ selectedTagId }: ClientPageProps) {
 
   if (loading) return <div>Loading...</div>;
   if (error) return <div>{error}</div>;
-
   return (
     <>
       <ArticleList
         articles={articles}
         sortedBy={sortBy}
         onSortChange={setSortBy}
+        onlyDrafts={onlyDrafts}
+        onlyArchived={onlyArchived}
+        onOnlyDraftsChange={setOnlyDrafts}
+        onOnlyArchivedChange={setOnlyArchived}
+        isAdmin={true}
       />
       {!lastPageReached && <div ref={loaderRef} className="h-10" />}
     </>
